fix: update perspective_projection to use f32 instead of f64 so T can be f32

+++ src/core/ops.rs
@@ -10,7 +10,7 @@ pub fn lerp<T>(u: T, v: T, coeff: f32) -> T where T: ops::Mul<f32, Output = T> +
 
 pub fn perspective_projection<T>(fov: T, aspect: T, near: T, far: T) -> Matrix<T>
 where
-    T: Field + Trig + From<f64>
+    T: Field + Trig + From<f32>
 {
     let mut mat = Matrix::<T>::from_elem(T::zero(), 4, 4);
     let inverse_tan_fov = T::one() / (fov / 2.0.into()).to_radians().tan();
@@ -19,7 +19,7 @@ where
     mat[(0, 0)] = inverse_aspect * inverse_tan_fov;
     mat[(1, 1)] = inverse_tan_fov;
     mat[(2, 2)] = -((far + near) / (far - near));
-    mat[(3, 2)] = -((<f64 as Into<T>>::into(2.0) * far * near) / (far - near));
+    mat[(3, 2)] = -((<f32 as Into<T>>::into(2.0) * far * near) / (far - near));
     mat[(2, 3)] = -T::one();
 
     mat
